+++ docs/suspensive.org/src/content/ko/docs/_meta.tsx
@@ -1,4 +1,5 @@
 import type { MetaRecord } from 'nextra'
+import type MetaEn from '../../en/docs/_meta'
 
 export default {
   '--- Packages': {
@@ -13,8 +14,19 @@ export default {
     type: 'separator',
     title: '더보기',
   },
-  'migrate-to-v2': 'v2로 마이그레이션하기',
-  changelogs: 'Changelogs',
-  contributors: '기여자',
-  links: '관련 링크',
-} satisfies MetaRecord
+  'migrate-to-v3': {
+    title: 'v3로 마이그레이션하기',
+  },
+  'migrate-to-v2': {
+    title: 'v2로 마이그레이션하기',
+  },
+  changelogs: {
+    title: 'Changelogs',
+  },
+  contributors: {
+    title: '기여자',
+  },
+  links: {
+    title: '관련 링크',
+  },
+} satisfies typeof MetaEn satisfies MetaRecord