+++ packages/react/src/Suspense.tsx
@@ -1,19 +1,9 @@
-import { type SuspenseProps as ReactSuspenseProps, useContext } from 'react'
+import { type ComponentProps, type ComponentType, type SuspenseProps as ReactSuspenseProps, useContext } from 'react'
 import { SuspenseDefaultPropsContext } from './contexts'
-import type { PropsWithDevMode } from './DevMode'
+import type { PropsWithoutChildren } from './utility-types'
 import { defineSuspense } from './utils'
 
-export interface SuspenseProps
-  extends PropsWithDevMode<
-    ReactSuspenseProps,
-    {
-      /**
-       * @deprecated Use official react devtools instead
-       * @see https://react.dev/learn/react-developer-tools
-       */
-      showFallback?: boolean
-    }
-  > {
+export interface SuspenseProps extends ReactSuspenseProps {
   /**
    * With clientOnly prop, `<Suspense/>` will return fallback in server but after mount return children in client. Since mount only happens on the client, `<Suspense/>` can be avoid server-side rendering.
    * @see https://suspensive.org/docs/react/Suspense#avoid-server-side-rendering-clientonly
@@ -25,17 +15,31 @@ export interface SuspenseProps
  * This component is just wrapping React's Suspense. to use Suspense easily in Server-side rendering environment like Next.js
  * @see {@link https://suspensive.org/docs/react/Suspense Suspensive Docs}
  */
-export const Suspense = ({ clientOnly, children, fallback }: SuspenseProps) => {
-  const defaultProps = useContext(SuspenseDefaultPropsContext)
-  const DefinedSuspense = defineSuspense({
-    defaultPropsClientOnly: defaultProps.clientOnly,
-    componentPropsClientOnly: clientOnly,
-  })
+export const Suspense = Object.assign(
+  ({ clientOnly, children, fallback }: SuspenseProps) => {
+    const defaultProps = useContext(SuspenseDefaultPropsContext)
+    const DefinedSuspense = defineSuspense({
+      defaultPropsClientOnly: defaultProps.clientOnly,
+      componentPropsClientOnly: clientOnly,
+    })
 
-  return (
-    <DefinedSuspense fallback={fallback === undefined ? defaultProps.fallback : fallback}>{children}</DefinedSuspense>
-  )
-}
-if (process.env.NODE_ENV === 'development') {
-  Suspense.displayName = 'Suspense'
-}
+    return (
+      <DefinedSuspense fallback={fallback === undefined ? defaultProps.fallback : fallback}>{children}</DefinedSuspense>
+    )
+  },
+  {
+    displayName: 'Suspense',
+    with: <TProps extends ComponentProps<ComponentType> = Record<string, never>>(
+      suspenseProps: PropsWithoutChildren<SuspenseProps> = {},
+      Component: ComponentType<TProps>
+    ) =>
+      Object.assign(
+        (props: TProps) => (
+          <Suspense {...suspenseProps}>
+            <Component {...props} />
+          </Suspense>
+        ),
+        { displayName: `Suspense.with(${Component.displayName || Component.name || 'Component'})` }
+      ),
+  }
+)