+++ packages/react/src/contexts/DefaultPropsContexts.ts
@@ -2,18 +2,12 @@ import { createContext } from 'react'
 import type { DelayProps, SuspenseProps } from '..'
 import type { OmitKeyof } from '../utility-types'
 
-export const DelayDefaultPropsContext = createContext<OmitKeyof<DelayProps, 'children'>>({
-  ms: undefined,
-  fallback: undefined,
-})
-if (process.env.NODE_ENV === 'development') {
-  DelayDefaultPropsContext.displayName = 'DelayDefaultPropsContext'
-}
+export const DelayDefaultPropsContext = Object.assign(
+  createContext<OmitKeyof<DelayProps, 'children'>>({ ms: undefined, fallback: undefined }),
+  { displayName: 'DelayDefaultPropsContext' }
+)
 
-export const SuspenseDefaultPropsContext = createContext<OmitKeyof<SuspenseProps, 'children'>>({
-  fallback: undefined,
-  clientOnly: undefined,
-})
-if (process.env.NODE_ENV === 'development') {
-  SuspenseDefaultPropsContext.displayName = 'SuspenseDefaultPropsContext'
-}
+export const SuspenseDefaultPropsContext = Object.assign(
+  createContext<OmitKeyof<SuspenseProps, 'children'>>({ fallback: undefined, clientOnly: undefined }),
+  { displayName: 'SuspenseDefaultPropsContext' }
+)