+++ packages/react/src/utility-types/index.ts
@@ -1,2 +1,3 @@
 export type { ConstructorType } from './ConstructorType'
 export type { OmitKeyof } from './OmitKeyof'
+export type { PropsWithoutChildren } from './PropsWithoutChildren'